requirements.txt
asgiref==3.4.1
Django==3.2.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
future==0.18.2
gunicorn==20.1.0
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.3.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
webencodings==0.5.1
